Top 500 similar users from the smaller dataset
This is the top 500 similar users from the small dataset consisting of 0,1,2,3,4,5 files. One thing I noticed was that a lot of the top people are just the same user. I wanted to at least have some output so we can start working on what to do now that we have the similarities. I got this just using $HADOOP_HOME/bin/hadoop fs -cat /final/calcSim/part-r-00000 | sort -n -k2 -r | head -n500. I changed the calcSimilarity package to use a DecimalFormatter so i could get the top using bash. I wrote a top mapreduce program but I was having errors. Will start on a mapreduce to compare the top users with relation to the time the site when down