Added mapreduce job to get activity of User
This gets the activity of the user based upon the file from MergeReducer. It finds how long the user was active, when the joined in relation to the first post, and the difference between their last post and the final post on the site. The format looks like this:
"Silk Road 2<==>yaoming<===>	160258<=A=>2489702<F=L>640246
The first number is time the user was active, aka LastPost - FirstPost, the delimiter <=A=> stands for active
The second number is time between the users first post and the database's first post, aka FirstPost - GlobalFirstPost 
The third number is the time between the last database post and the users last post, aka GlobalLastPost - LastPost.
The delimiter <F=L> stands for first on the left side and last on the left side.